luxen_backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
firebase-admin==6.2.0
google-cloud-firestore==2.13.0
google-cloud-storage==2.10.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.107.0
numpy==1.24.3
scikit-learn==1.3.2
requests==2.31.0
gunicorn==21.2.0
python-dateutil==2.8.2

